arm64: Use per-cpu page tables
Same as on x86: Set up per-cpu tables, mapping the per-cpu structure
at a fixed location. This allows to drop per-cpu logic based on
TPIDR_EL2.

We have to ensure that not only the common mappings of hv_paging_structs
are linked but also that ID mapping of the trampoline page is working.
It is needed during shutdown. Furthermore, we need to take care that
arch_shutdown_self() is called with a pointer into the common per-cpu
mapping so that paging_hvirt2phys() works.

The stack pointer is left on the common mapping. It will be switched
when leaving to EL1.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>