x86: Convert from per_cpu address calculation to multiplication
As long as the sizeof(struct per_cpu) is a power of two, the compiler
actually does the right thing for us. Whenever it should no longer be
power of two, we deviate from the driver that simply does cpu * size to
calculate the per-cpu area size. So let's drop the fragile assembly
optimization.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>