arm-common: Move pending_irqs into public per-cpu section
All fields of this structure are used across CPUs in order to maintain
the software IRQ queue, thus need to be public.

This encourages parameter type changes for irqchip_cpu_shutdown and
irqchip_set_pending. Both functions only operate on the public per-cpu
section now, which should be reflected in the API. irqchip_cpu_shutdown
may now be called when private per-cpu sections are hidden without
worrying of the pointer is locally or commonly mapped.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>